moved configure-projects script back to root
